ml/requirements.txt
fastapi
uvicorn
catboost
pandas
scikit-learn
numpy
requests
python-dotenv
pydantic
shap
optuna
